optimize saltstack to keep uniform style
