+++ plugins/mysql_user.go
@@ -1,0 +1,154 @@
+package plugins
+
+import (
+	"fmt"
+
+	"github.com/sirupsen/logrus"
+)
+
+var MysqlDatabaseUserPluginActions = make(map[string]Action)
+
+func init() {
+	MysqlDatabaseUserPluginActions["addUser"] = new(AddMysqlDatabaseUserAction)
+}
+
+type MysqlUserPlugin struct {
+}
+
+func (plugin *MysqlUserPlugin) GetActionByName(actionName string) (Action, error) {
+	action, found := MysqlDatabaseUserPluginActions[actionName]
+	if !found {
+		return nil, fmt.Errorf("mysql user plugin,action = %s not found", actionName)
+	}
+
+	return action, nil
+}
+
+//------------AddMysqlDatabaseUserAction--------------
+type AddMysqlDatabaseUserAction struct {
+}
+
+type AddMysqlDatabaseUserInputs struct {
+	Inputs []AddMysqlDatabaseUserInput `json:"inputs,omitempty"`
+}
+
+type AddMysqlDatabaseUserInput struct {
+	CallBackParameter
+	Guid     string `json:"guid,omitempty"`
+	Seed     string `json:"seed,omitempty"`
+	Host     string `json:"host,omitempty"`
+	UserName string `json:"userName,omitempty"`
+	Password string `json:"password,omitempty"`
+	Port     string `json:"port,omitempty"`
+	// AccessKey string `json:"accessKey,omitempty"`
+	// SecretKey string `json:"secretKey,omitempty"`
+
+	//new database info
+	DatabaseUserGuid     string `json:"databaseUserGuid,omitempty"`
+	DatabaseName         string `json:"databaseName,omitempty"`
+	DatabaseUserName     string `json:"databaseUserName,omitempty"`
+	DatabaseUserPassword string `json:"databaseUserPassword,omitempty"`
+}
+
+type AddMysqlDatabaseUserOutputs struct {
+	Outputs []AddMysqlDatabaseUserOutput `json:"outputs,omitempty"`
+}
+
+type AddMysqlDatabaseUserOutput struct {
+	CallBackParameter
+	Result
+	DatabaseUserGuid     string `json:"databaseUserGuid,omitempty"`
+	DatabaseUserPassword string `json:"databaseUserPassword,omitempty"`
+}
+
+func (action *AddMysqlDatabaseUserAction) ReadParam(param interface{}) (interface{}, error) {
+	var inputs AddMysqlDatabaseUserInputs
+	if err := UnmarshalJson(param, &inputs); err != nil {
+		return nil, err
+	}
+	return inputs, nil
+}
+
+func checkAddMysqlDatabaseUser(input *AddMysqlDatabaseUserInput) error {
+	if input.Guid == "" {
+		return fmt.Errorf("empty guid")
+	}
+	if input.Seed == "" {
+		return fmt.Errorf("empty seed")
+	}
+	if input.Password == "" {
+		return fmt.Errorf("empty password")
+	}
+	if input.DatabaseUserName == "" {
+		return fmt.Errorf("empty databaseUserName")
+	}
+	if input.DatabaseUserGuid == "" {
+		return fmt.Errorf("empty databaseUserGuid")
+	}
+	return nil
+}
+
+func createUserForExistedDatabase(input *AddMysqlDatabaseUserInput) (string, error) {
+	if err := checkAddMysqlDatabaseUser(input); err != nil {
+		return "", err
+	}
+
+	//get root password
+	password, err := AesDePassword(input.Guid, input.Seed, input.Password)
+	if err != nil {
+		logrus.Errorf("AesDePassword meet error(%v)", err)
+		return "", err
+	}
+
+	//create user
+	userPassword := input.DatabaseUserPassword
+	if userPassword == "" {
+		userPassword = createRandomPassword()
+	}
+
+	cmd := fmt.Sprintf("CREATE USER %s IDENTIFIED BY '%s' ", input.DatabaseUserName, userPassword)
+	if err = runDatabaseCommand(input.Host, input.Port, input.UserName, password, cmd); err != nil {
+		return "", err
+	}
+
+	//grant permission
+	if input.DatabaseName != "" {
+		permission := "ALL PRIVILEGES"
+		cmd = fmt.Sprintf("GRANT %s ON %s.* TO %s ", permission, input.DatabaseName, input.DatabaseUserName)
+		if err = runDatabaseCommand(input.Host, input.Port, input.UserName, password, cmd); err != nil {
+			return "", err
+		}
+	}
+
+	//create encrypt password
+	encryptPassword, err := AesEnPassword(input.Guid, input.Seed, userPassword, DEFALT_CIPHER)
+	if err != nil {
+		logrus.Errorf("AesEnPassword meet error(%v)", err)
+		return "", err
+	}
+	return encryptPassword, err
+}
+
+func (action *AddMysqlDatabaseUserAction) Do(input interface{}) (interface{}, error) {
+	inputs, _ := input.(AddMysqlDatabaseUserInputs)
+	outputs := AddMysqlDatabaseUserOutputs{}
+	var finalErr error
+
+	for _, input := range inputs.Inputs {
+		output := AddMysqlDatabaseUserOutput{
+			DatabaseUserGuid: input.DatabaseUserGuid,
+		}
+		output.CallBackParameter.Parameter = input.CallBackParameter.Parameter
+		output.Result.Code = RESULT_CODE_SUCCESS
+
+		password, err := createUserForExistedDatabase(&input)
+		if err != nil {
+			finalErr = err
+			output.Result.Code = RESULT_CODE_ERROR
+			output.Result.Message = err.Error()
+		}
+		output.DatabaseUserPassword = password
+		outputs.Outputs = append(outputs.Outputs, output)
+	}
+	return outputs, finalErr
+}
